chore!: restructure project
```
├── CHANGELOG.md
├── Cargo.lock
├── Cargo.toml
├── LICENSE
├── README.md
└── src
    ├── lib.rs
    ├── lib_tests.rs
    └── m3u8
        ├── mod.rs
        ├── parser.rs
        ├── playlist
        │ ├── builder.rs
        │   └── mod.rs
        ├── tags.rs
        └── validation.rs

```
