+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### ⚙️ Miscellaneous Tasks
 
+- [**breaking**] Restructure codes
+
+## [0.4.0] - 2024-11-01
+
+### ⚙️ Miscellaneous Tasks
+
 - Add a Note about the status of the project
 - Fix `Crates.io` badge in `README.md`
