+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "m3u8-parser"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Amin Jamal <aminjamal10@gmail.com>"]
 edition = "2021"
 license = "MIT"
