+++ src/m3u8/mod.rs
@@ -1,0 +1,5 @@
+pub mod parser;
+pub mod playlist;
+pub mod tags;
+mod tests;
+pub mod validation;
