+++ src/m3u8/tests/mod.rs
@@ -1,0 +1,1 @@
+mod lib_tests;
